refactor(#1401): SonarCloud CPD 51줄 dup 제거 — 두 fusion 테스트 파일 병합
- useFusedNearestStation.forwardOnly.test.ts (222줄) → trainProgressing.test.ts (368줄)에 통합
- jest.mock 9개 블록 + import + cast 약 50줄을 단일 파일에 1회만 선언 (Jest 호이스팅 제약 회피)
- fixture tp-(boarding idx=0) / fwd-(boarding idx=1) 둘 다 const 유지
- sonar.cpd.exclusions에서 삭제된 파일 경로 제거

sonar.cpd.exclusions에 명시했지만 SonarCloud가 dup 6.2% > 3% 게이트 fail. 글로브 파싱 한계로 추정 → 물리 병합으로 robust 해결.
