+++ sonar-project.properties
@@ -14,4 +14,4 @@ sonar.exclusions=assets/**,**/__tests__/**,**/__tests__/**/*,src/**/__tests__/**
 # Live Activity ActivityKit 구조체가 widget 타겟과 CocoaPod 모듈에서 동일 정의를
 # 요구하는 의도된 중복(_shared 자동 링크가 pod 컴파일 단위 커버 못 함)이라 불가피.
 # i18n locales는 다국어 번역으로 키 구조 동일이 정상 — CPD 의미 없음.
-sonar.cpd.exclusions=**/*.swift,**/SubwayActivityAttributes.swift,targets/**,targets/**/*,modules/**,modules/**/*,src/shared/i18n/locales/**,**/__tests__/**,**/__tests__/**/*,**/*.test.ts,**/*.test.tsx,src/**/__tests__/**,src/**/__tests__/**/*,app/**/__tests__/**,src/data/timetables/**,**/__tests__/*.ts,**/__tests__/*.tsx,backend/**,backend/**/*,src/features/nearest-station/utils/fusionDebugBuffer.ts,src/features/route/utils/estimatorDebugBuffer.ts,src/shared/utils/createDebugBuffer.ts,src/features/nearest-station/utils/fusionDistanceGate.ts,src/testUtils/**,src/features/alarm/utils/pushConsistency.ts,src/features/alarm/utils/__tests__/pushConsistency.test.ts,src/features/alarm/utils/__tests__/pushConsistency.mirror.test.ts,backend/alarm-worker/src/__tests__/pushConsistency.test.ts,backend/alarm-worker/src/pushConsistency.ts,src/features/nearest-station/hooks/__tests__/useFusedNearestStation.trainProgressing.test.ts,src/features/nearest-station/hooks/__tests__/useFusedNearestStation.forwardOnly.test.ts,src/features/alarm/hooks/__tests__/useStationAlarm.test.ts,src/features/nearest-station/utils/__tests__/movementGate.test.ts
+sonar.cpd.exclusions=**/*.swift,**/SubwayActivityAttributes.swift,targets/**,targets/**/*,modules/**,modules/**/*,src/shared/i18n/locales/**,**/__tests__/**,**/__tests__/**/*,**/*.test.ts,**/*.test.tsx,src/**/__tests__/**,src/**/__tests__/**/*,app/**/__tests__/**,src/data/timetables/**,**/__tests__/*.ts,**/__tests__/*.tsx,backend/**,backend/**/*,src/features/nearest-station/utils/fusionDebugBuffer.ts,src/features/route/utils/estimatorDebugBuffer.ts,src/shared/utils/createDebugBuffer.ts,src/features/nearest-station/utils/fusionDistanceGate.ts,src/testUtils/**,src/features/alarm/utils/pushConsistency.ts,src/features/alarm/utils/__tests__/pushConsistency.test.ts,src/features/alarm/utils/__tests__/pushConsistency.mirror.test.ts,backend/alarm-worker/src/__tests__/pushConsistency.test.ts,backend/alarm-worker/src/pushConsistency.ts,src/features/nearest-station/hooks/__tests__/useFusedNearestStation.trainProgressing.test.ts,src/features/alarm/hooks/__tests__/useStationAlarm.test.ts,src/features/nearest-station/utils/__tests__/movementGate.test.ts
